codestream: move to klplib package
